+++ main.py
@@ -123,6 +123,7 @@ def app_layout():
             dcc.Store(id="memory-submitted-data", storage_type="memory"),  # Rinkmenų kortelėje patvirtinti duomenys
             dcc.Store(id="memory-selected-tables", storage_type="session"),  # Pasirinktos lentelės (be kaimynų)
             dcc.Store(id="memory-filtered-data", storage_type="memory"),   # Grafiko piešimui atrinkti duomenys
+            dcc.Store(id="viz-key-press-store", data=""),  # žr. assets/main.js; neveikia kaip pastovi atmintis
             dcc.Store(id="viz-clicked-node-store", data=""),  # žr. assets/main.js; neveikia kaip pastovi atmintis
             dcc.Store(id="viz-clicked-checkbox-store", data=""),  # žr. assets/main.js; neveikia kaip pastovi atmintis
             dcc.Store(id="memory-last-selected-nodes", storage_type="memory"),  # žr. get_selected_node_ids()